Move git commands to git directory
Changing the path of git commands from src/renderer/mixins/git to src/renderer/git
