+++ src/renderer/git/commit.js
