+++ src/renderer/git/diff.js
