+++ src/renderer/git/show.js
